moved add_notes_to_total into the correct mod no longer associated function
